Some work on GL 3.2 / GLSL 1.50 support

+++ executables/sandbox/main/update.cpp
@@ -32,7 +32,7 @@ void game::update()
    }
 
    m_debug_renderer->clear_text_lines();
-   m_debug_renderer->record_frame_duration(m_frame_dt);
+   m_debug_renderer->record_frame_duration(static_cast<float>(m_frame_dt));
 
    //if (!m_paused)
    {
